extract some primitives to parachain crate
